implement generic message sending class
